Move CachePolicyFactory tests to tests/v1/kv_offload/cpu/policies/test_factory.py
Mirrors the tiering/factory.py <-> tiering/test_factory.py convention,
per review feedback on PR #49114.
